+++ internal/provider/provider.go
@@ -181,6 +181,8 @@ func (p *EmporixProvider) Configure(ctx context.Context, req provider.ConfigureR
 func (p *EmporixProvider) Resources(ctx context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
 		NewSiteSettingsResource,
+		NewPaymentModeResource,
+		NewCountryResource,
 	}
 }